requirements.txt
PyYAML==6.0.1
Random_Word==1.0.11
streamlit==1.26.0
streamlit_authenticator==0.2.3
streamlit_extras==0.4.3
streamlit_survey==0.1.0
tqdm==4.66.1